+++ R/mod2values.R
@@ -89,6 +89,8 @@ mod2values <- function(x){
     }
     gnames <- rep(names(PrepList), each = length(est)/length(PrepList))
     par[parname %in% c('g', 'u')] <- antilogit(par[parname %in% c('g', 'u')])
+    lbound[parname %in% c('g', 'u')] <- antilogit(lbound[parname %in% c('g', 'u')])
+    ubound[parname %in% c('g', 'u')] <- antilogit(ubound[parname %in% c('g', 'u')])
     prior.type <- sapply(as.character(prior.type),
                          function(x) switch(x, '1'='norm', '2'='lnorm', '3'='beta', 'none'))
     ret <- data.frame(group=gnames, item=item, class=class, name=parname, parnum=parnum, value=par,