+++ R/model.elements.R
@@ -127,6 +127,68 @@ model.elements <- function(model, factorNames, itemtype, nfactNames, nfact, J, K
     ret[[length(ret) + 1L]] <- LoadGroupPars(gmeans=gmeans, gcov=gcov, estgmeans=estgmeans,
                                             estgcov=estgcov, parnumber=attr(ret, 'parnumber'),
                                             parprior=parprior, Rasch=all(itemtype %in% c('Rasch', 'rsm')))
+    if(any(model[,1L] == 'LBOUND')){
+        input <- gsub(" ","", model[model[,1L] == 'LBOUND', 2L])
+        elements <- strsplit(input, '\\),\\(')[[1L]]
+        elements <- gsub('\\(', replacement='', x=elements)
+        elements <- gsub('\\)', replacement='', x=elements)
+        esplit <- strsplit(elements, ',')
+        esplit <- lapply(esplit, function(x){
+            newx <- c()
+            if(length(x) < 3L)
+                stop('LBOUND = ... has not been supplied enough arguments', call.=FALSE)
+            for(i in 1L:(length(x)-2L)){
+                if(grepl('-', x[i])){
+                    tmp <- as.numeric(strsplit(x[i], '-')[[1L]])
+                    newx <- c(newx, tmp[1L]:tmp[2L])
+                } else newx <- c(newx, x[i])
+            }
+            x <- c(newx, x[length(x)-1L], x[length(x)])
+            x
+        })
+        picks <- sapply(esplit, function(x) as.integer(x[1L:(length(x)-2)]))
+        for(i in 1L:length(picks)){
+            tmp <- ret[picks[[i]]]
+            len <- length(esplit[[i]])
+            tmp <- lapply(tmp, function(x, which, val){
+                if(which %in% c('g', 'u')) val <- qlogis(val)
+                x@lbound[names(x@parnum) == which] <- val
+                x
+            }, which=esplit[[i]][len-1L], val = as.numeric(esplit[[i]][len]))
+            ret[picks[[i]]] <- tmp
+        }
+    }
+    if(any(model[,1L] == 'UBOUND')){
+        input <- gsub(" ","", model[model[,1L] == 'LBOUND', 2L])
+        elements <- strsplit(input, '\\),\\(')[[1L]]
+        elements <- gsub('\\(', replacement='', x=elements)
+        elements <- gsub('\\)', replacement='', x=elements)
+        esplit <- strsplit(elements, ',')
+        esplit <- lapply(esplit, function(x){
+            newx <- c()
+            if(length(x) < 3L)
+                stop('UBOUND = ... has not been supplied enough arguments', call.=FALSE)
+            for(i in 1L:(length(x)-2L)){
+                if(grepl('-', x[i])){
+                    tmp <- as.numeric(strsplit(x[i], '-')[[1L]])
+                    newx <- c(newx, tmp[1L]:tmp[2L])
+                } else newx <- c(newx, x[i])
+            }
+            x <- c(newx, x[length(x)-1L], x[length(x)])
+            x
+        })
+        picks <- sapply(esplit, function(x) as.integer(x[1L:(length(x)-2)]))
+        for(i in 1L:length(picks)){
+            tmp <- ret[picks[[i]]]
+            len <- length(esplit[[i]])
+            tmp <- lapply(tmp, function(x, which, val){
+                if(which %in% c('g', 'u')) val <- qlogis(val)
+                x@lbound[names(x@parnum) == which] <- val
+                x
+            }, which=esplit[[i]][len-1L], val = as.numeric(esplit[[i]][len]))
+            ret[picks[[i]]] <- tmp
+        }
+    }
     attr(ret, 'prodlist') <- prodlist
     attr(ret, 'exploratory') <- exploratory
     return(ret)